requirements.txt
fastapi>=0.104.0
uvicorn>=0.24.0
scikit-learn>=1.3.0
joblib>=1.3.0
pandas>=2.0.0
numpy>=1.24.0
python-dotenv>=1.0.0
supabase>=2.3.0
streamlit>=1.28.0
matplotlib>=3.7.0
seaborn>=0.12.0
requests>=2.31.0
ollama>=0.1.0
openai>=1.0.0
